Leet Code DS first commit
