+++ data-structures/src/main/java/com/design/analysis/leetcode/ds/single/linkedlist/medium/ISLinkedListMedium.java
@@ -1,8 +1,8 @@
-package com.design.analysis.leetcode.ds.single.linkedlist;
+package com.design.analysis.leetcode.ds.single.linkedlist.medium;
 
 import com.design.analysis.leetcode.ds.node.SLNode;
 
-public interface ISLinkedList<K> {
+public interface ISLinkedListMedium<K> {
 
     /**2. Add Two Numbers, The digits are stored in reverse order, and each of their nodes contains a single digit**/
     SLNode<K> addTwoNumbers(SLNode<K> l1, SLNode<K> l2);
@@ -11,7 +11,4 @@ public interface ISLinkedList<K> {
 
     /**19. Remove Nth Node From End of List*/
     SLNode<K> removeNthFromEnd(SLNode<K> l1, int n);
-
-    /**21. Merge Two Sorted Lists, space efficient, not use extra space**/
-    SLNode<K> mergeTwoLists(SLNode<K> l1,SLNode<K> l2);
 }
