+++ data-structures/src/main/java/com/design/analysis/leetcode/ds/stack/easy/IStackEasy.java
@@ -1,0 +1,7 @@
+package com.design.analysis.leetcode.ds.stack.easy;
+
+public interface IStackEasy {
+
+    /** 20. Valid Parentheses*/
+    boolean isValidParentheses(String s);
+}
